pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
authors = [{ name = "Mikhail Kokuev"}]
description = "Task manager"
requires-python = ">=3.14"
dependencies = [
    "dj-database-url>=3.1.2",
    "django>=6.0.4",
    "django-bootstrap5>=26.2",
    "django-filter>=25.2",
    "django-widget-tweaks>=1.5.1",
    "gunicorn>=25.3.0",
    "psycopg2-binary>=2.9.12",
    "python-decouple>=3.8",
]

[dependency-groups]
dev = [
    "coverage>=7.13.5",
    "ruff>=0.15.12",
]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
preview = true
select = ["E", "F", "I", "C90"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
